added output for citations
To record which citations have been used in the impact statements, we created a list of citations (as we find them). Once we encounter the bibliography at the end of the paper, we then loop through each unique citation to see if any of them were used in the impact statement. This is then compiled into another dataframe (and another CSV output) for reference (database style).