separated notebook into scrapping and analysis
For easier organization (and tracking) I took out the "scrapping" functions of the script (i.e. take XML files and output CSV) and placed it into a proper python file. The reason for this is for simplified change tracking (since we don't really need this portion to be in a Jupyter notebook).

I left the analysis portion (currently incomplete) in the notebook (and renamed it), since that one benefits from the UI of a Jupyter notebook. However, changes will need to be made to the notebook to pull from an external output.csv file instead of an internal dictionary.