Modified script to account for multiple entries in a section
Modified the code significantly to account for multiple entries for the conclusion. I did this by establishing a concatenating for-loop. Once the script identifies an adequate header it will loop through all subsequent entries checking for the conditional. If the condition is satisfied it will concatenate the text contents into impact_statement_text.

The contents of impact_statement_text are then consolidated and processed outside the for-loop once the entire document has been parsed.

I also made some minor changes to account for the h1-h2 scenario.

Should close #10